+++ PlexMatchGenerator/Helpers/ArgumentHelper.cs
@@ -6,7 +6,16 @@ namespace PlexMatchGenerator.Helpers
 {
     public class ArgumentHelper
     {
-        public static GeneratorOptions ProcessCommandLineResults(string plexToken, string plexUrl, List<string> rootPaths, string logPath, bool noOverwrite, int pageSize)
+        public static GeneratorOptions ProcessCommandLineResults(
+            string plexToken, 
+            string plexUrl, 
+            List<string> rootPaths, 
+            string logPath, 
+            bool noOverwrite, 
+            int pageSize, 
+            List<string> libraries, 
+            List<string> shows,
+            bool seasonProcessing)
         {
             //ensure we end the path with a slash
             if (logPath != null && !logPath.EndsWith("\\") && !logPath.EndsWith('/'))
@@ -24,7 +33,10 @@ namespace PlexMatchGenerator.Helpers
                 PlexServerToken = plexToken,
                 RootPaths = GenerateRootPaths(rootPaths),
                 NoOverwrite = noOverwrite,
-                ItemsPerPage = pageSize == 0 ? 20 : pageSize
+                ItemsPerPage = pageSize == 0 ? 20 : pageSize,
+                LibraryNames = libraries,
+                ShowNames = shows,
+                EnablePerSeasonProcessing = seasonProcessing
             };
         }