+++ .github/workflows/ReleaseNotes.md
@@ -1,6 +1,11 @@
+# Release Notes
+
 ## What's Changed
-* Add command switch to disable automatic overwriting as requested in [#30](https://github.com/johnkiddjr/PlexMatch-File-Generator/issues/30)
-* Added paging support to help relieve sever stress as requested in [#6](https://github.com/johnkiddjr/PlexMatch-File-Generator/issues/6)
+
+* Add library command line switch to restrict to only specified libraries per [#38](https://github.com/johnkiddjr/PlexMatch-File-Generator/issues/38)
+* Add show command line switch to restrict to only specified media items per [#37](https://github.com/johnkiddjr/PlexMatch-File-Generator/issues/37)
+* Add seasonprocessing command line switch to enable per season processing of TV Shows per [#39](https://github.com/johnkiddjr/PlexMatch-File-Generator/issues/39)
 
 ## Known Bugs
+
 * None! If you encounter any bugs, please open an [issue](https://github.com/johnkiddjr/PlexMatch-File-Generator/issues/new)!